System.ZMQ3.Monadic

Description

Type Definitions

data ZMQ a

class (Monad m, MonadIO m, MonadCatchIO m, MonadPlus m, Functor m, Applicative m, Alternative m) => MonadZMQ m where

Methods

liftZMQ :: ZMQ a -> m a

data Socket a

A 0MQ Socket.

data Flag

Flags to apply on send operations (cf. man zmq_send)

Constructors

SendMore
ZMQ_SNDMORE

data Switch

Configuration switch

Constructors

Default
Use default setting

On
Activate setting

Off
De-activate setting

type Timeout = Int64

data Event

Socket events.

Constructors

In
ZMQ_POLLIN (incoming messages)

Out
ZMQ_POLLOUT (outgoing messages, i.e. at least 1 byte can be written)

Err
ZMQ_POLLERR

data EventType

Event types to monitor.

Constructors

ConnectedEvent

ConnectDelayedEvent

ConnectRetriedEvent

ListeningEvent

BindFailedEvent

AcceptedEvent

AcceptFailedEvent

ClosedEvent

CloseFailedEvent

DisconnectedEvent

AllEvents

data EventMsg

Event Message to receive when monitoring socket events.

Constructors

Connected !ByteString !Fd

ConnectDelayed !ByteString !Fd

ConnectRetried !ByteString !Int

Listening !ByteString !Fd

BindFailed !ByteString !Fd

Accepted !ByteString !Fd

AcceptFailed !ByteString !Int

Closed !ByteString !Fd

CloseFailed !ByteString !Int

Disconnected !ByteString !Int

data Poll m where

Type representing a descriptor, poll is waiting for (either a 0MQ socket or a file descriptor) plus the type of event to wait for.

Constructors

Sock :: Socket s -> [Event] -> Maybe ([Event] -> m ()) -> Poll m

File :: Fd -> [Event] -> Maybe ([Event] -> m ()) -> Poll m

Type Classes

class SocketType a

Socket types.

class Sender a

Sockets which can send.

class Receiver a

Sockets which can receive.

class Subscriber a

Sockets which can subscribe.

Socket Types

data Pair

Socket to communicate with a single peer. Allows for only a single connect or a single bind. There's no message routing or message filtering involved. Compatible peer sockets: Pair.

Constructors

Pair

data Pub

Socket to distribute data. receive function is not implemented for this socket type. Messages are distributed in fanout fashion to all the peers. Compatible peer sockets: Sub.

Constructors

Pub

data Sub

Socket to subscribe for data. send function is not implemented for this socket type. Initially, socket is subscribed for no messages. Use subscribe to specify which messages to subscribe for. Compatible peer sockets: Pub.

Constructors

Sub

data XPub

Same as Pub except that you can receive subscriptions from the peers in form of incoming messages. Subscription message is a byte 1 (for subscriptions) or byte 0 (for unsubscriptions) followed by the subscription body. Compatible peer sockets: Sub, XSub.

Constructors

XPub

data XSub

Same as Sub except that you subscribe by sending subscription messages to the socket. Subscription message is a byte 1 (for subscriptions) or byte 0 (for unsubscriptions) followed by the subscription body. Compatible peer sockets: Pub, XPub.

Constructors

XSub

data Req

Socket to send requests and receive replies. Requests are load-balanced among all the peers. This socket type allows only an alternated sequence of send's and recv's. Compatible peer sockets: Rep, Router.

Constructors

Req

data Rep

Socket to receive requests and send replies. This socket type allows only an alternated sequence of receive's and send's. Each send is routed to the peer that issued the last received request. Compatible peer sockets: Req, Dealer.

Constructors

Rep

data Dealer

Each message sent is round-robined among all connected peers, and each message received is fair-queued from all connected peers. Compatible peer sockets: Router, Req, Rep.

Constructors

Dealer

data Router

When receiving messages a Router socket shall prepend a message part containing the identity of the originating peer to the message before passing it to the application. Messages received are fair-queued from among all connected peers. When sending messages a Router socket shall remove the first part of the message and use it to determine the identity of the peer the message shall be routed to. If the peer does not exist anymore the message shall be silently discarded. Compatible peer sockets: Dealer, Req, Rep.

Constructors

Router

data Pull

A socket of type Pull is used by a pipeline node to receive messages from upstream pipeline nodes. Messages are fair-queued from among all connected upstream nodes. The zmq_send() function is not implemented for this socket type.

Constructors

Pull

data Push

A socket of type Push is used by a pipeline node to send messages to downstream pipeline nodes. Messages are load-balanced to all connected downstream nodes. The zmq_recv() function is not implemented for this socket type.

When a Push socket enters an exceptional state due to having reached the high water mark for all downstream nodes, or if there are no downstream nodes at all, then any zmq_send(3) operations on the socket shall block until the exceptional state ends or at least one downstream node becomes available for sending; messages are not discarded.

Constructors

Push

General Operations

version :: MonadZMQ m => m (Int, Int, Int)

runZMQ :: MonadIO m => ZMQ a -> m a

async :: ZMQ a -> ZMQ ()

socket :: (MonadZMQ m, SocketType t) => t -> m (Socket t)

ZMQ Options (Read)

ioThreads :: MonadZMQ m => m Word

maxSockets :: MonadZMQ m => m Word

ZMQ Options (Write)

setIoThreads :: MonadZMQ m => Word -> m ()

setMaxSockets :: MonadZMQ m => Word -> m ()

Socket operations

close :: MonadZMQ m => Socket t -> m ()

bind :: MonadZMQ m => Socket t -> String -> m ()

unbind :: MonadZMQ m => Socket t -> String -> m ()

connect :: MonadZMQ m => Socket t -> String -> m ()

send :: (MonadZMQ m, Sender t) => Socket t -> [Flag] -> ByteString -> m ()

send' :: (MonadZMQ m, Sender t) => Socket t -> [Flag] -> ByteString -> m ()

sendMulti :: (MonadZMQ m, Sender t) => Socket t -> [ByteString] -> m ()

receive :: (MonadZMQ m, Receiver t) => Socket t -> m ByteString

receiveMulti :: (MonadZMQ m, Receiver t) => Socket t -> m [ByteString]

subscribe :: (MonadZMQ m, Subscriber t) => Socket t -> ByteString -> m ()

unsubscribe :: (MonadZMQ m, Subscriber t) => Socket t -> ByteString -> m ()

proxy :: MonadZMQ m => Socket a -> Socket b -> Maybe (Socket c) -> m ()

monitor :: MonadZMQ m => [EventType] -> Socket t -> m (Bool -> IO (Maybe EventMsg))

poll :: MonadIO m => Timeout -> [Poll m] -> m [[Event]]

Polls for events on the given Poll descriptors. Returns the same list of Poll descriptors with an updated PollEvent field (cf. zmq_poll). Sockets which have seen no activity have None in their PollEvent field.

Socket Options (Read)

affinity :: MonadZMQ m => Socket t -> m Word64

backlog :: MonadZMQ m => Socket t -> m Int

delayAttachOnConnect :: MonadZMQ m => Socket t -> m Bool

events :: MonadZMQ m => Socket t -> m [Event]

fileDescriptor :: MonadZMQ m => Socket t -> m Fd

identity :: MonadZMQ m => Socket t -> m ByteString

ipv4Only :: MonadZMQ m => Socket t -> m Bool

lastEndpoint :: MonadZMQ m => Socket t -> m String

linger :: MonadZMQ m => Socket t -> m Int

maxMessageSize :: MonadZMQ m => Socket t -> m Int64

mcastHops :: MonadZMQ m => Socket t -> m Int

moreToReceive :: MonadZMQ m => Socket t -> m Bool

rate :: MonadZMQ m => Socket t -> m Int

receiveBuffer :: MonadZMQ m => Socket t -> m Int

receiveHighWM :: MonadZMQ m => Socket t -> m Int

receiveTimeout :: MonadZMQ m => Socket t -> m Int

reconnectInterval :: MonadZMQ m => Socket t -> m Int

reconnectIntervalMax :: MonadZMQ m => Socket t -> m Int

recoveryInterval :: MonadZMQ m => Socket t -> m Int

sendBuffer :: MonadZMQ m => Socket t -> m Int

sendHighWM :: MonadZMQ m => Socket t -> m Int

sendTimeout :: MonadZMQ m => Socket t -> m Int

tcpKeepAlive :: MonadZMQ m => Socket t -> m Switch

tcpKeepAliveCount :: MonadZMQ m => Socket t -> m Int

tcpKeepAliveIdle :: MonadZMQ m => Socket t -> m Int

tcpKeepAliveInterval :: MonadZMQ m => Socket t -> m Int

Socket Options (Write)

setAffinity :: MonadZMQ m => Word64 -> Socket t -> m ()

setBacklog :: (MonadZMQ m, Integral i) => Restricted N0 Int32 i -> Socket t -> m ()

setDelayAttachOnConnect :: MonadZMQ m => Bool -> Socket t -> m ()

setIdentity :: MonadZMQ m => Restricted N1 N254 ByteString -> Socket t -> m ()

setIpv4Only :: MonadZMQ m => Bool -> Socket t -> m ()

setLinger :: (MonadZMQ m, Integral i) => Restricted Nneg1 Int32 i -> Socket t -> m ()

setMaxMessageSize :: (MonadZMQ m, Integral i) => Restricted Nneg1 Int64 i -> Socket t -> m ()

setMcastHops :: (MonadZMQ m, Integral i) => Restricted N1 Int32 i -> Socket t -> m ()

setRate :: (MonadZMQ m, Integral i) => Restricted N1 Int32 i -> Socket t -> m ()

setReceiveBuffer :: (MonadZMQ m, Integral i) => Restricted N0 Int32 i -> Socket t -> m ()

setReceiveHighWM :: (MonadZMQ m, Integral i) => Restricted N0 Int32 i -> Socket t -> m ()

setReceiveTimeout :: (MonadZMQ m, Integral i) => Restricted Nneg1 Int32 i -> Socket t -> m ()

setReconnectInterval :: (MonadZMQ m, Integral i) => Restricted N0 Int32 i -> Socket t -> m ()

setReconnectIntervalMax :: (MonadZMQ m, Integral i) => Restricted N0 Int32 i -> Socket t -> m ()

setRecoveryInterval :: (MonadZMQ m, Integral i) => Restricted N0 Int32 i -> Socket t -> m ()

setRouterMandatory :: MonadZMQ m => Bool -> Socket Router -> m ()

setSendBuffer :: (MonadZMQ m, Integral i) => Restricted N0 Int32 i -> Socket t -> m ()

setSendHighWM :: (MonadZMQ m, Integral i) => Restricted N0 Int32 i -> Socket t -> m ()

setSendTimeout :: (MonadZMQ m, Integral i) => Restricted Nneg1 Int32 i -> Socket t -> m ()

setTcpAcceptFilter :: MonadZMQ m => Maybe ByteString -> Socket t -> m ()

setTcpKeepAlive :: MonadZMQ m => Switch -> Socket t -> m ()

setTcpKeepAliveCount :: (MonadZMQ m, Integral i) => Restricted Nneg1 Int32 i -> Socket t -> m ()

setTcpKeepAliveIdle :: (MonadZMQ m, Integral i) => Restricted Nneg1 Int32 i -> Socket t -> m ()

setTcpKeepAliveInterval :: (MonadZMQ m, Integral i) => Restricted Nneg1 Int32 i -> Socket t -> m ()

setXPubVerbose :: MonadZMQ m => Bool -> Socket XPub -> m ()

Error Handling

data ZMQError

ZMQError encapsulates information about errors, which occur when using the native 0MQ API, such as error number and message.

errno :: ZMQError -> Int

Error number value.

source :: ZMQError -> String

Source where this error originates from.

message :: ZMQError -> String

Actual error message.

Re-exports

liftIO :: MonadIO m => forall a. IO a -> m a

Lift a computation from the IO monad.

restrict :: Restriction l u v => v -> Restricted l u v

Create a restricted value. If the given value does not satisfy the restrictions, a modified variant is used instead, e.g. if an integer is larger than the upper bound, the upper bound value is used.

toRestricted :: Restriction l u v => v -> Maybe (Restricted l u v)

Create a restricted value. Returns Nothing if the given value does not satisfy all restrictions.

Low-level Functions

waitRead :: MonadZMQ m => Socket t -> m ()

waitWrite :: MonadZMQ m => Socket t -> m ()